+++ xmake/modules/private/action/require/impl/install_packages.lua
@@ -217,12 +217,26 @@ function _get_confirm(packages, opt)
                         local group = instance:group()
                         if group and packages_group[group] and #packages_group[group] > 1 then
                             for idx, package_in_group in ipairs(packages_group[group]) do
-                                cprint("  ${yellow}%s${clear} %s %s ${dim}%s", idx == 1 and "->" or "   or", package_in_group:displayname(), package_in_group:version_str() or "", package.get_configs_str(package_in_group))
+                                cprint("  ${yellow}%s${clear} %s %s ${dim}%s", idx == 1 and "->" or "   or",
+                                    package_in_group:displayname(), package_in_group:version_str() or "",
+                                    package.get_configs_str(package_in_group))
+                                for _, tip in ipairs(package_in_group:get("installtips")) do
+                                    if idx == 1 then
+                                        cprint("     ${yellow}*${clear} %s", tip)
+                                    else
+                                        cprint("        ${yellow}*${clear} %s", tip)
+                                    end
+                                end
                                 packages_showed[tostring(package_in_group)] = true
                             end
                             packages_group[group] = nil
                         else
-                            cprint("  ${yellow}->${clear} %s %s ${dim}%s", instance:displayname(), instance:version_str() or "", package.get_configs_str(instance))
+                            cprint("  ${yellow}->${clear} %s %s ${dim}%s",
+                                instance:displayname(), instance:version_str() or "",
+                                package.get_configs_str(instance))
+                            for _, tip in ipairs(instance:get("installtips")) do
+                                cprint("     ${yellow}*${clear} %s", tip)
+                            end
                             packages_showed[tostring(instance)] = true
                         end
                     end
@@ -731,7 +745,12 @@ function _install_packages(requires, opt)
     if #packages_unknown > 0 then
         cprint("${bright color.warning}note: ${clear}the following packages were not found in any repository (check if they are spelled correctly):")
         for _, instance in ipairs(packages_unknown) do
-            print("  -> %s", instance:displayname())
+            local tips
+            local possible_package = package.get_possible_package(instance:name())
+            if possible_package then
+                tips = string.format(", maybe ${bright}%s %s${clear} in %s", possible_package.name, possible_package.version, possible_package.reponame)
+            end
+            cprint("  -> %s%s", instance:displayname(), tips or "")
         end
         has_errors = true
     end