+++ xmake/modules/private/action/require/impl/package.lua
@@ -32,6 +32,7 @@ import("core.platform.platform")
 import("core.package.package", {alias = "core_package"})
 import("devel.git")
 import("private.action.require.impl.repository")
+import("private.action.require.impl.search_packages")
 import("private.action.require.impl.utils.requirekey", {alias = "_get_requirekey"})
 
 -- get memcache
@@ -138,7 +139,8 @@ function _parse_require(require_str)
 end
 
 -- load require info
-function _load_require(require_str, requires_extra, parentinfo)
+function _load_require(require_str, requires_extra, opt)
+    opt = opt or {}
 
     -- parse require
     local packagename, version, reponame = _parse_require(require_str)
@@ -181,6 +183,26 @@ function _load_require(require_str, requires_extra, parentinfo)
         end
     end
 
+    -- resolve require info
+    local resolvedinfo = opt.resolvedinfo
+    local requirepath = opt.requirepath
+    if requirepath then
+        requirepath = requirepath .. "." .. packagename
+    else
+        requirepath = packagename
+    end
+    resolvedinfo = resolvedinfo and resolvedinfo[requirepath]
+    if resolvedinfo then
+        -- resolve the conflict package version
+        if resolvedinfo.version then
+            version = resolvedinfo.version
+        end
+        -- resolve the conflict package configs
+        if resolvedinfo.configs then
+            require_extra.configs = resolvedinfo.configs
+        end
+    end
+
     -- get required building configurations
     -- we need to clone a new configs object, because the whole requireinfo will be modified later.
     -- @see https://github.com/xmake-io/xmake-repo/pull/2067
@@ -216,6 +238,7 @@ function _load_require(require_str, requires_extra, parentinfo)
 
     -- init required item
     local required = {}
+    local parentinfo = opt.parentinfo
     parentinfo = parentinfo or {}
     required.packagename = packagename
     required.requireinfo =
@@ -238,7 +261,8 @@ function _load_require(require_str, requires_extra, parentinfo)
         verify           = require_extra.verify,    -- default: true, we can set false to ignore sha256sum and select any version
         external         = require_extra.external,  -- default: true, we use sysincludedirs/-isystem instead of -I/xxx
         private          = require_extra.private,   -- default: false, private package, only for installation, do not export any links/includes and environments
-        build            = require_extra.build      -- default: false, always build packages, we do not use the precompiled artifacts
+        build            = require_extra.build,     -- default: false, always build packages, we do not use the precompiled artifacts
+        resolvedinfo     = resolvedinfo             -- the resolved info for the conflict version/configs
     }
     return required.packagename, required.requireinfo
 end
@@ -967,8 +991,17 @@ function _load_package(packagename, requireinfo, opt)
         package = _load_package_from_system(packagename)
     end
 
-    -- check
-    assert(package, "package(%s) not found!", packagename)
+    -- check unknown package
+    if not package then
+        cprint("${bright color.warning}note: ${clear}the following packages were not found in any repository (check if they are spelled correctly):")
+        local tips
+        local possible_package = get_possible_package(packagename)
+        if possible_package then
+            tips = string.format(", maybe ${bright}%s %s${clear} in %s", possible_package.name, possible_package.version, possible_package.reponame)
+        end
+        cprint("  -> %s%s", packagename, tips or "")
+        raise("package(%s) not found!", packagename)
+    end
 
     -- init requireinfo
     _init_requireinfo(requireinfo, package, {is_toplevel = not opt.parentinfo})
@@ -1001,6 +1034,7 @@ function _load_package(packagename, requireinfo, opt)
     local version, source = _select_package_version(package, requireinfo, locked_requireinfo)
     if version then
         package:version_set(version, source)
+        package:data_set("__locked_requireinfo", locked_requireinfo)
     end
 
     -- get package key
@@ -1110,7 +1144,7 @@ function _load_packages(requires, opt)
         -- load package
         local requireinfo = requireitem.info
         local requirepath = opt.requirepath and (opt.requirepath .. "." .. requireitem.name) or requireitem.name
-        local package     = _load_package(requireitem.name, requireinfo, table.join(opt, {requirepath = requirepath}))
+        local package = _load_package(requireitem.name, requireinfo, table.join(opt, {requirepath = requirepath}))
 
         -- maybe package not found and optional
         if package then
@@ -1124,6 +1158,7 @@ function _load_packages(requires, opt)
                                                         requires_extra = package:extraconf("deps") or {},
                                                         parentinfo = requireinfo,
                                                         nodeps = opt.nodeps,
+                                                        resolvedinfo = opt.resolvedinfo,
                                                         system = false})
                     for _, dep in ipairs(plaindeps) do
                         dep:parents_add(package)
@@ -1166,7 +1201,199 @@ function _get_parents_str(package)
     end
 end
 
--- check dependencies conflicts
+-- get require paths
+function _get_requirepaths(package)
+    local requirepaths = {}
+    local parents = package:parents()
+    if parents and #parents > 0 then
+        for _, parent in ipairs(parents) do
+            for _, requirepath in ipairs(_get_requirepaths(parent)) do
+                table.insert(requirepaths, requirepath .. "." .. package:name())
+            end
+        end
+        -- we also need to resolve requires conflict in toplevel, if `package.sync_requires_to_deps` policy is enabled.
+        -- @see https://github.com/xmake-io/xmake/issues/5745#issuecomment-2513951471
+        if project.policy("package.sync_requires_to_deps") then
+            table.insert(requirepaths, package:name())
+        end
+    else
+        table.insert(requirepaths, package:name())
+    end
+    return requirepaths
+end
+
+-- get compatibility key
+function _get_package_compatkey(dep)
+    local key = dep:plat() .. "/" .. dep:arch() .. "/" .. (dep:kind() or "")
+    if dep:is_system() then
+        key = key .. "/system"
+    end
+    local resolve_depconflict = project.policy("package.resolve_depconflict")
+    if resolve_depconflict == nil then
+        resolve_depconflict = dep:policy("package.resolve_depconflict")
+    end
+    if resolve_depconflict == false then
+        if dep:version_str() then
+            key = key .. "/" .. dep:version_str()
+        end
+        local configs = dep:requireinfo().configs
+        if configs then
+            local configs_order = {}
+            for k, v in pairs(configs) do
+                if type(v) == "table" then
+                    v = string.serialize(v, {strip = true, indent = false, orderkeys = true})
+                end
+                table.insert(configs_order, k .. "=" .. tostring(v))
+            end
+            table.sort(configs_order)
+            key = key .. ":" .. string.serialize(configs_order, true)
+        end
+
+    end
+    return key
+end
+
+-- get package compatibility info
+function _get_package_compatinfo(package)
+    local compatinfo = {name = package:name()}
+    if package:data("__locked_requireinfo") then
+        compatinfo.locked = true
+        compatinfo.versions = hashset.from(table.wrap(package:version_str()))
+        return compatinfo
+    end
+
+    -- get compatible version range
+    local requireinfo = package:requireinfo()
+    local require_version = requireinfo.version
+    if require_version then
+        compatinfo.require_version = require_version
+        if semver.is_valid(require_version) or semver.is_valid_range(require_version) then
+            local versions = hashset.new()
+            for _, version in ipairs(package:versions()) do
+                if semver.satisfies(version, require_version) then
+                    versions:insert(version)
+                end
+            end
+            compatinfo.versions = versions
+        elseif require_version == "latest" then
+            compatinfo.versions = hashset.from(table.wrap(package:versions()))
+        else
+            compatinfo.versions = hashset.from(table.wrap(require_version))
+        end
+    else
+        compatinfo.versions = hashset.from(table.wrap(package:versions()))
+    end
+    return compatinfo
+end
+
+-- check and resolve package conflicts
+function _check_and_resolve_package_depconflicts_impl(package, name, deps, resolvedinfo)
+
+    -- check version compatibility
+    local versions
+    for _, dep in ipairs(deps) do
+        local compatinfo = _get_package_compatinfo(dep)
+        assert(compatinfo.versions)
+
+        if versions then
+            for version in versions:items() do
+                if not compatinfo.versions:has(version) then
+                    versions:remove(version)
+                end
+            end
+        else
+            versions = compatinfo.versions
+        end
+    end
+
+    if not versions or versions:empty() then
+        print("package(%s): add_deps(%s, ...)", package:name(), name)
+        for idx, dep in ipairs(deps) do
+            cprint("  ${color.warning}->${clear} %s %s ${dim}%s",
+                dep:displayname(), dep:version_str() or "", get_configs_str(dep))
+        end
+        print("we can use add_requireconfs(\"**.%s\", {override = true, version = \"x.x.x\"}) to override version.", name)
+        raise("package(%s): conflict version dependencies!", name)
+    end
+
+    -- check configs compatibility
+    local prevkey
+    local configs
+    local configs_conflict = false
+    for _, dep in ipairs(deps) do
+        local key = _get_package_compatkey(dep)
+        if prevkey then
+            if prevkey ~= key then
+                configs_conflict = true
+                break
+            end
+        else
+            prevkey = key
+        end
+        local depconfigs = dep:requireinfo().configs
+        if configs and depconfigs then
+            for k, v in pairs(depconfigs) do
+                local oldv = configs[k]
+                if oldv ~= nil then
+                    local v_key = tostring(v)
+                    local oldv_key = tostring(oldv)
+                    if type(v) == "table" then
+                        v_key = string.serialize(v, {strip = true, indent = false, orderkeys = true})
+                    end
+                    if type(oldv) == "table" then
+                        oldv_key = string.serialize(oldv, {strip = true, indent = false, orderkeys = true})
+                    end
+                    if v_key ~= oldv_key then
+                        configs_conflict = true
+                        break
+                    end
+                else
+                    configs[k] = v
+                end
+            end
+        else
+            configs = depconfigs
+        end
+    end
+    if configs_conflict then
+        print("package(%s): add_deps(%s, ...)", package:name(), name)
+        for idx, dep in ipairs(deps) do
+            cprint("  ${color.warning}->${clear} %s %s ${dim}%s",
+                dep:displayname(), dep:version_str() or "", get_configs_str(dep))
+        end
+        print("we can use add_requireconfs(\"**.%s\", {override = true, configs = {}}) to override configs.", name)
+        raise("package(%s): conflict configs dependencies!", name)
+    end
+
+    -- resolve compatible version for all deps
+    if versions and not versions:empty() then
+        local version_best
+        for version in versions:items() do
+            if version_best == nil or semver.compare(version, version_best) > 0 then
+                version_best = version
+            end
+        end
+        if version_best then
+            for _, dep in ipairs(deps) do
+                for _, requirepath in ipairs(_get_requirepaths(dep)) do
+                    resolvedinfo[requirepath] = {version = version_best}
+                end
+            end
+        end
+    end
+
+    -- resolve configs
+    if configs then
+        for _, dep in ipairs(deps) do
+            for _, requirepath in ipairs(_get_requirepaths(dep)) do
+                resolvedinfo[requirepath] = resolvedinfo[requirepath] or {}
+                resolvedinfo[requirepath].configs = configs
+            end
+        end
+    end
+end
+
+-- check and resolve dependencies conflicts
 --
 -- It exists conflict for dependent packages for each root packages? resolve it first
 -- e.g.
@@ -1180,15 +1407,19 @@ end
 -- Of course, conflicts caused by `add_packages("foo", "ddd")`
 -- cannot be detected at present and can only be resolved by the user
 --
-function _check_package_depconflicts(package)
-    local packagekeys = {}
+function _check_and_resolve_package_depconflicts(package, resolvedinfo)
+    local some_packagedeps = {}
     for _, dep in ipairs(package:librarydeps()) do
-        local key = _get_packagekey(dep:name(), dep:requireinfo())
-        local prevkey = packagekeys[dep:name()]
-        if prevkey then
-            assert(key == prevkey, "package(%s): conflict dependencies with package(%s) in %s!", key, prevkey, package:name())
-        else
-            packagekeys[dep:name()] = key
+        local deps = some_packagedeps[dep:name()]
+        if deps == nil then
+            deps = {}
+            some_packagedeps[dep:name()] = deps
+        end
+        table.insert(deps, dep)
+    end
+    for name, deps in pairs(some_packagedeps) do
+        if #deps > 1 then
+            _check_and_resolve_package_depconflicts_impl(package, name, deps, resolvedinfo)
         end
     end
 end
@@ -1282,6 +1513,24 @@ function _compatible_with_previous_librarydeps(package, opt)
         package:data_set("force_reinstall", true)
     end
     return is_compatible
+end
+
+-- get possible package
+function get_possible_package(packagename)
+    local packages_possible = search_packages("*", {description = false})
+    if packages_possible then
+        packages_possible = packages_possible["*"]
+    end
+    local result
+    local distance_min
+    for name, info in pairs(packages_possible) do
+        local distance = packagename:levenshtein(info.name)
+        if distance_min == nil or distance < distance_min and distance < 5 then
+            distance_min = distance
+            result = info
+        end
+    end
+    return result
 end
 
 -- the cache directory
@@ -1380,10 +1629,18 @@ function get_configs_str(package)
     if parents_str then
         table.insert(configs, "from:" .. parents_str)
     end
+    local license = package:get("license")
+    if license then
+        table.insert(configs, "license:" .. license)
+    end
     local configs_str = #configs > 0 and "[" .. table.concat(configs, ", ") .. "]" or ""
     local limitwidth = math.floor(os.getwinsize().width * 2 / 3)
     if #configs_str > limitwidth then
         configs_str = configs_str:sub(1, limitwidth) .. " ..)"
+    end
+    local resolvedinfo = requireinfo.resolvedinfo
+    if resolvedinfo then
+        configs_str = configs_str .. " " .. "${color.warning}(conflict resolved)${clear}"
     end
     return configs_str
 end
@@ -1406,7 +1663,7 @@ function load_requires(requires, requires_extra, opt)
     opt = opt or {}
     local requireitems = {}
     for _, require_str in ipairs(requires) do
-        local packagename, requireinfo = _load_require(require_str, requires_extra, opt.parentinfo)
+        local packagename, requireinfo = _load_require(require_str, requires_extra, opt)
         table.insert(requireitems, {name = packagename, info = requireinfo})
     end
     return requireitems
@@ -1417,14 +1674,36 @@ function load_packages(requires, opt)
     opt = opt or {}
     local unique = {}
     local packages = {}
-    for _, package in ipairs((_load_packages(requires, opt))) do
+    local resolvedinfo = {}
+    for _, package in ipairs((_load_packages(requires, table.clone(opt)))) do
         if package:is_toplevel() then
-            _check_package_depconflicts(package)
+            _check_and_resolve_package_depconflicts(package, resolvedinfo)
         end
         local key = _get_packagekey(package:name(), package:requireinfo())
         if not unique[key] then
             table.insert(packages, package)
             unique[key] = true
+        end
+    end
+
+    -- we need to reload packages with new resolved deps if there are some dep conflicts
+    if not table.empty(resolvedinfo) then
+        for _, package in ipairs(packages) do
+            local installdir = package:installdir({readonly = true})
+            if os.isdir(installdir) and os.emptydir(installdir) then
+                os.tryrm(installdir, {emptydirs = true})
+            end
+        end
+        unique = {}
+        packages = {}
+        _memcache():clear()
+        opt = table.join(opt, {resolvedinfo = resolvedinfo})
+        for _, package in ipairs((_load_packages(requires, opt))) do
+            local key = _get_packagekey(package:name(), package:requireinfo())
+            if not unique[key] then
+                table.insert(packages, package)
+                unique[key] = true
+            end
         end
     end
     return packages